% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_functions.R
\name{dag_funcs}
\alias{dag_funcs}
\alias{.common_ancestors}
\alias{.set_edge_weight}
\alias{.longest_path}
\title{GAPGOM internal - Directed acyclic graph (DAG) mathematical functions/DAG 
navigation functions.}
\usage{
.common_ancestors(go_id1, go_id2, ontology, organism, go_annotation,
  information_content)

.set_edge_weight(sub_graph, information_content)

.longest_path(weighted_subgraph, last_common_ancestor, root,
  information_content)
}
\description{
These functions are internal functions and should not be called by the user.
}
\details{
Mathematical calculations and navigations functions for directed acyclic 
graphs (DAG). These functions do a multitude of things; calculate longest
paths between two nodes, calculate edge weight and retrieving a common 
ancestor in the GO DAG given two GO ids.
}
\section{Notes}{

Internal functions all used in topo_ic_sim_titj().
}

