% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.entrezraw_to_entrez}
\alias{.entrezraw_to_entrez}
\title{GAPGOM internal - .entrezraw_to_entrez()}
\usage{
.entrezraw_to_entrez(rawid)
}
\arguments{
\item{rawid}{Raw entrez id. E.G.; entrez:23456}
}
\value{
return the normal ID
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Converts a raw ID from fantom data to a normal ID.
}
\section{Notes}{

Internal function used in enrichment_analysis().
}

\keyword{internal}
