% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.ext_id_to_term_id}
\alias{.ext_id_to_term_id}
\title{GAPGOM internal - ext_id_to_term_id()}
\usage{
.ext_id_to_term_id(data, list_top_genes)
}
\arguments{
\item{data}{extracted_genes with correct onto etc. (matrix)}

\item{list_top_genes}{The list_top_genes matrix}
}
\value{
return the quantified matrix
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Convert an ext ID to a term ID dat = EG_ (Extracted genes with correct gene ontology and narrowed selection of top genes.)  list_top_genes (List of top 250 genes)
}
\section{Notes}{

Internal function used in enrichment_analysis().
}

