% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.generate_translation_df}
\alias{.generate_translation_df}
\title{GAPGOM internal - .generate_translation_df()}
\usage{
.generate_translation_df(expression_set, organism, ontology, keytype,
  verbose = F, go_data = NULL)
}
\arguments{
\item{expression_set}{ExpressionSet object --> see Biobase package.}

\item{organism}{where to be scanned genes reside in, this option
is neccesary to select the correct GO DAG. Options are based on the org.db
bioconductor package;
http://www.bioconductor.org/packages/release/BiocViews.html#___OrgDb
Following options are available: "fly", "mouse", "rat", "yeast",
"zebrafish", "worm", "arabidopsis", "ecolik12", "bovine", "canine",
"anopheles", "ecsakai", "chicken", "chimp", "malaria", "rhesus", "pig",
"xenopus". Fantom5 data only has "human" and "mouse" available depending
on the dataset.}

\item{ontology}{desired ontology to use for prediction. One of three;
"BP" (Biological process), "MF" (Molecular function) or "CC"
(Cellular Component). Cellular Component is not included with the package's
standard data and will thus yield no results.}

\item{keytype}{keytype used in querying of godata/columnnames}

\item{verbose}{set to true for more informative/elaborate output.}
}
\value{
return the translation dataframe containing conversion from general 
ids to entrez/ensembl ids (and others) and goids.
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Generates a translation df specific for ExpressionSets, this is then used
for looking up ids and their respective GOIDS.
}
\section{Notes}{

Internal function used in expression_prediction_function().
}

\keyword{internal}
