% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.get_package_version}
\alias{.get_package_version}
\title{GAPGOM internal - .get_package_version()}
\usage{
.get_package_version(pckg_name)
}
\arguments{
\item{pckg_name}{package name}
}
\value{
package version
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Gets the version number of a package.
}
\section{Notes}{

Internal function used in checking precalculated values for topoicsim.
}

\keyword{internal}
