% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.go_ids_lookup}
\alias{.go_ids_lookup}
\title{GAPGOM internal - .go_ids_lookup()}
\usage{
.go_ids_lookup(ids, go_data, custom_genes = NULL, drop = NULL)
}
\arguments{
\item{ids}{general ids that you want to search for in godata.}

\item{go_data}{the queried godata neccesary for the lookup}

\item{drop}{list of evidences you want to ignore.}
}
\value{
return the translation dataframe containing conversion from id to 
goids.
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Looks up goids per id (entrez/ensembl).
}
\section{Notes}{

Internal function used in ().
}

\keyword{internal}
