% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.organism_to_species_lib}
\alias{.organism_to_species_lib}
\title{GAPGOM internal - .organism_to_species_lib()}
\usage{
.organism_to_species_lib(organism)
}
\arguments{
\item{organism}{the desired organism}
}
\value{
the lib name
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Makes conversion of organism names to package names a little easier
}
\section{Notes}{

Internal function used in set_go_data().
}

\keyword{internal}
