% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.prepare_variables_topoicsim}
\alias{.prepare_variables_topoicsim}
\title{GAPGOM internal - .prepare_variables_topoicsim()}
\usage{
.prepare_variables_topoicsim(organism, ontology, gene_list1 = NULL,
  gene_list2 = NULL, drop = NULL, verbose = F, progress_bar = NULL,
  garbage_collection = NULL, all_go_pairs = NULL, topoargs = list(),
  term_only = FALSE, keytype = "ENTREZID")
}
\arguments{
\item{organism}{where to be scanned genes reside in, this option
is neccesary to select the correct GO DAG. Options are based on the org.db
bioconductor package;
http://www.bioconductor.org/packages/release/BiocViews.html#___OrgDb
Following options are available: "fly", "mouse", "rat", "yeast",
"zebrafish", "worm", "arabidopsis", "ecolik12", "bovine", "canine",
"anopheles", "ecsakai", "chicken", "chimp", "malaria", "rhesus", "pig",
"xenopus".}

\item{ontology}{desired ontology to use for similarity calculations.
One of three;
"BP" (Biological process),
"MF" (Molecular function) or
"CC" (Cellular Component).}

\item{gene_list1}{The first gene vector of gene IDs. Note; THIS IS NOT THE
ENSEMBLID. Instead use the gene ID adopted by NCBI.}

\item{gene_list2}{Same type as gene_list1, will be compared to gene_list1.}

\item{drop}{vector of GOID you want to exclude from the analysis.}

\item{verbose}{set to true for more informative/elaborate output.}

\item{progress_bar}{Whether to show the progress of the calculation 
(default = FALSE)}

\item{garbage_collection}{whether to do R garbage collection. This is
useful for very large calculations/datasets, as it might decrease ram usage.
This option might however increase calculation time.}

\item{all_go_pairs}{dataframe of GO Term pairs with a column
representing similarity between the two. You can add the dataframe from
previous runs to improve performance (only works if the last result has
at least part of the genes of the current run). You can also use it for 
pre-calculation and getting the results back in a fast manner.}

\item{topoargs}{topoargs list that needs correction, default is a brand new
list}

\item{term_only}{specify if you only want arguments/params for TERM level.}

\item{keytype}{keytype used in querying of godata}
}
\value{
list with all topoicsim arguments (can differ depending on 
algorithm level gene/geneset/term); organism, ontology, verbose, IC 
(Information Content from the go_data/go consortium), weighted_dag (DAG 
with weighted nodes), go_annotation (GO annotation of the correct root 
node), root (root node of the GO tree (MF, BP or CC) in 
annotationdbi-database string form), drop, progress_bar, garbage_collection, 
all_go_pairs, selected_freq_go_pairs (precalculated common GO_pairs that 
might increase performance), translation_to_goids (translation dataframe 
between ID and GOID.)
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Prepares a list of intermediary arguments/parameters needed by topocisim.
}
\section{Notes}{

Internal function used in (topo_ic_sim_genes).
}

\keyword{internal}
