% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.resolve_entrezid_col}
\alias{.resolve_entrezid_col}
\title{GAPGOM internal - .resolve_entrez_col()}
\usage{
.resolve_entrezid_col(expression_set)
}
\arguments{
\item{expression_set}{ExpressionSet object --> see Biobase package.}
}
\value{
column name of entrez id
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
resolves columnname for entrez ids within an expressoin_set
}
\section{Notes}{

Internal function used in .generate_translation_df().
}

