% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.resolve_genes_unique_gos}
\alias{.resolve_genes_unique_gos}
\title{GAPGOM internal - .resolve_genes_unique_gos()}
\usage{
.resolve_genes_unique_gos(unique_pairs_genes, topoargs)
}
\arguments{
\item{unique_pairs_genes}{data.table with unique pairs of gene combinations.}

\item{topoargs}{topoicsim args, see data_prep}
}
\value{
return a data.table with all unique go combination pairs.
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Resolves all unique go pairs given unique gene pairs.
}
\section{Notes}{

Internal function used in topo_ic_sim_genes().
}

\keyword{internal}
