% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.resolve_keys_col}
\alias{.resolve_keys_col}
\title{GAPGOM internal - .resolve_keys_col()}
\usage{
.resolve_keys_col(expression_set, keytype)
}
\arguments{
\item{expression_set}{ExpressionSet object --> see Biobase package.}

\item{keytype}{keytype used in querying of godata}
}
\value{
column name of id
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
resolves columnname for arbitrary ids within an expression_set
}
\section{Notes}{

Internal function used in .generate_translation_df().
}

\keyword{internal}
