% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.set_values}
\alias{.set_values}
\title{GAPGOM internal - .set_values()}
\usage{
.set_values(item1, item2, the_matrix, value)
}
\arguments{
\item{item1}{first item}

\item{item2}{second item}

\item{the_matrix}{score matrix}

\item{value}{value to be set}
}
\value{
return the matrix with newly set items.
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Sets values within a topoclsim score matrix.
}
\section{Notes}{

Internal function used in expression_prediction_function().
}

\keyword{internal}
