% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{.term_id_to_ext_id}
\alias{.term_id_to_ext_id}
\title{GAPGOM internal - .term_id_to_ext_id()}
\usage{
.term_id_to_ext_id(data)
}
\arguments{
\item{data}{extracted_genes with correct onto etc. (matrix)}
}
\value{
return the quantified matrix
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Quantifies amount of total extracted go terms
}
\section{Notes}{

Internal function used in enrichment_analysis().
}

\keyword{internal}
