% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_checks.R
\name{.topo_ic_sim_argcheck_genes}
\alias{.topo_ic_sim_argcheck_genes}
\title{GAPGOM internal - .topo_ic_sim_argchecks_genes()}
\usage{
.topo_ic_sim_argcheck_genes(ontology, organism, genes1, genes2)
}
\value{
no return value, stops code when arguments are faulty.
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
This function checks arguments for the topoicsim gene variant. This function
is temporary and will be replaced with other checks later on.
}
\keyword{internal}
