% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{.topo_ic_sim_g1g2}
\alias{.topo_ic_sim_g1g2}
\title{GAPGOM internal - .topo_ic_sim_g1g2()}
\usage{
.topo_ic_sim_g1g2(gene1, gene2, topoargs)
}
\arguments{
\item{gene1}{Gene ID of the first Gene.}

\item{gene2}{Gene ID of the second Gene.}

\item{topoargs}{list containing all the neccesary paramters/arguments for
the topoicsim algorithm, details can be view in the 
".prepare_variables_topoicsim" function.}
}
\value{
List containing the following;
$GeneSim;
similarity between genes taken from the mean of all term 
similarities.
$AllGoPairs;
All possible GO combinations with their semantic distances (matrix)
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Algorithm to calculate similarity between GO terms of two genes.
This function is made for calculating topological similarity of two Genes
given their GO terms in the GO DAG structure. The topological similarity is
based on edge weights and information content (IC). [1]
}
\section{Notes}{

Internal function used in (topo_ic_sim_genes()).
}

\references{
[1] Ehsani R, Drablos F: \strong{TopoICSim: a new semantic
similarity measure based on gene ontology.} \emph{BMC Bioinformatics} 2016,
\strong{17}(1):296)
}
