% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{.topo_ic_sim_geneset}
\alias{.topo_ic_sim_geneset}
\title{GAPGOM internal - .topo_ic_sim()}
\usage{
.topo_ic_sim_geneset(gene_list1, gene_list2, topoargs)
}
\arguments{
\item{gene_list1}{The first gene vector of gene IDs. Note; THIS IS NOT THE
ENSEMBLID. Instead use the gene ID adopted by NCBI.}

\item{gene_list2}{Same type as gene_list1, will be compared to gene_list1.}

\item{topoargs}{list containing all the neccesary paramters/arguments for
the topoicsim algorithm, details can be view in the 
".prepare_variables_topoicsim" function.}
}
\value{
List containing the following;
$GeneSim;
similarity between genes taken from the mean of all term 
similarities within those genes. (matrix of similarities between genes).
Take the mean of this matrix to either get the InterSetSim or IntraSetSim
depending on your input.
$AllGoPairs;
All possible GO combinations with their semantic distances (matrix)
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Algorithm to calculate similarity between two gene vectors.
This function is made for calculating topological similarity between two
gene lists of which each gene has its GO terms in the GO DAG structure.
The topological similarity is based on edge weights and information
content (IC). The output it a nxn matrix depending on the vector lengths.
Intraset similarity can be calculated by comparing the same gene vector to
itself and using mean() on the output. The same can be done for Interset
similarity, but between two \strong{different} gene lists. [1]
}
\references{
[1] Ehsani R, Drablos F: \strong{TopoICSim: a new semantic
similarity measure based on gene ontology.} \emph{BMC Bioinformatics} 2016,
\strong{17}(1):296)
}
