% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{.topo_ic_sim_titj}
\alias{.topo_ic_sim_titj}
\title{GAPGOM internal - .topo_ic_sim_titj()}
\usage{
.topo_ic_sim_titj(go_id1, go_id2, topoargs)
}
\arguments{
\item{go_id1}{GO term of first term.}

\item{go_id2}{GO term of second term.}

\item{topoargs}{list containing all the neccesary paramters/arguments for
the topoicsim algorithm, details can be view in the 
".prepare_variables_topoicsim" function.}
}
\value{
semantic topological similarity between the two given go terms.
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Algorithm to calculate similarity between two GO terms.
This function is made for calculating topological similarity of two GO
terms in the GO DAG structure. The topological similarity is based on
edge weights and information content (IC). [1]
}
\references{
[1] Ehsani R, Drablos F: \strong{TopoICSim: a new semantic
similarity measure based on gene ontology.} \emph{BMC Bioinformatics} 2016,
\strong{17}(1):296)
}
