% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{.unique_combos}
\alias{.unique_combos}
\title{GAPGOM internal - .unique_combos()}
\usage{
.unique_combos(v1, v2)
}
\arguments{
\item{v1}{first vector to be combined with second.}

\item{v2}{second vector to be combined with first.}
}
\value{
datatable with unique combinations of the vectors excluding items
with same name and double combinations
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Generates unique combinations of two vectors
}
\section{Notes}{

Internal function used in various topoclsim algorithms.
}

