% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_analysis}
\alias{enrichment_analysis}
\title{GAPGOM internal - enrichment_analysis()}
\usage{
enrichment_analysis(ordered_score_df, ontology,
  expression_data = GAPGOM::expression_data,
  ensembl_to_go_id_conversion_df = GAPGOM::ensembl_id_to_go_id,
  enrichment_cutoff = 250)
}
\arguments{
\item{ordered_score_df}{the score dataframe see documentation on
GAPGOM::example_score_dataframe for formatting.}

\item{ontology}{desired ontology to use for prediction. One of three;
"BP" (Biological process), "MF" (Molecular function) or "CC"
(Cellular Component). Cellular Component is not included with the package's
standard data and will thus yield no results.}

\item{expression_data}{dataframe with at least 1 or more
rows. Some of the columns should be named certain ways and the dataframe
should contain certain information neccesary for calculation;
1st neccesary col; name; "GeneID" (EnsemblIDs should go here).
2nd neccesary col; name; "GeneType" (genetype as described in the Ensembl
database).
Next, a range of n-columns long, containing expression values, names do not
matter here. (You can use the existing dataset as an example)}

\item{ensembl_to_go_id_conversion_df}{dataframe with corresponding GOIDs and
EnsemblIDs. As well as the expression value dataframe, this dataframe should
contain certain information with specific columnnames as well;
1st neccesary col; name; "ensembl_gene_id" (Ensembl gene IDs).
2nd neccesary col; go_id; "go_id" (Gene Ontology term ID).
3rd and last neccesary col; "namespace_1003" (Toplevel GO term).
(You can use the existing dataset as an example)}

\item{enrichment_cutoff}{cutoff number for the amount of genes to be
enriched in the enrichment analysis. (default is 250)}
}
\value{
The resulting dataframe with prediction of similar GO terms.
These are ordered with respect to FDR values. The following columns will be
in the dataframe;
GOID - Gene Ontology ID,
Ontology - Ontology type (MF or BP),
FDR - False Positive Rate,
Term - description of GOID.
However, unlike in expression_prediction, this dataframe will have unsorted
row numbering. And it won't contain used method.
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Enriches score results from multiple methods to give a better idea of
important similarities.
This function is specifically made for predicting lncRNA annotation by
assuming "guilt by association". For instance, the expression data in this
package is actually based on mRNA expression data, but correlated with
lncRNA. This expression data is the used in combination with mRNA GO
annotation to calculate similarity scores between GO terms,
}
\section{Notes}{

Internal function used in expression_prediction_function().
}

