% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_id_to_go_id.R
\docType{data}
\name{ensembl_id_to_go_id}
\alias{ensembl_id_to_go_id}
\title{EnsemblID -> GOID conversion table}
\format{A data frame with 252887 rows and 3 variables/columns:
\describe{
  \item{ensembl_gene_id}{The EnsemblID of the gene.}
  \item{go_id}{The GOID of the property. Multiple GOIDs per EnsemblIDs
  are possible.}
  \item{namespace_1003}{Toplevel GO node.}
}}
\usage{
ensembl_id_to_go_id
}
\description{
A dataframe containing EnsemblIDs and corresponding GOIDs for refering back
and forth between the IDs.
}
\details{
Source unknown.
}
\keyword{datasets}
