% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_data.R
\docType{data}
\name{expression_data}
\alias{expression_data}
\title{GAPGOM - Expression data of lncRNAs.}
\format{A data frame with 33164 rows and 22 variables/columns:
\describe{
  \item{GeneID}{EnsemblID of gene}
  \item{GeneType}{Type of gene. E.g. 'protein_coding'}
  \item{GeneName}{Standard abbreviated gene name as it recognized in most
  databases}
  \item{x_FPKM}{Gene expression in FPKM (Fragments Per Kilobase Million)
  where x is a type of tissue (e.g. placenta).}
  ...
}}
\source{
\url{http://mlg.hit.edu.cn:80/lncrna2function/download.jsp}
}
\usage{
expression_data
}
\description{
A dataframe containing lncRNA expression values that have been correlated
with mRNA expression. [1]
}
\details{
The host url/website is unavailable.
}
\references{
[1]. Jiang Q, Ma R, Wang J, Wu X, Jin S, Peng J, Tan R, Zhang T,
Li Y, Wang Y: \strong{LncRNA2Function: a comprehensive resource for
functional investigation of human lncRNAs based on RNA-seq data.}
\emph{BMC Genomics 2015}, \strong{16 Suppl 3}:S2.
}
\keyword{datasets}
