% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_prediction.R
\name{expression_prediction_function}
\alias{expression_prediction_function}
\title{GAPGOM - expression_prediction_function()}
\usage{
expression_prediction_function(gene_id, ontology,
  expression_data = GAPGOM::expression_data,
  ensembl_to_go_id_conversion_df = GAPGOM::ensembl_id_to_go_id,
  start_expression_col = 4, end_expression_col = 22,
  enrichment_cutoff = 250, method = "combine")
}
\arguments{
\item{gene_id}{gene EnsemblID to be compared to the other GO terms.}

\item{ontology}{desired ontology to use for prediction. One of three;
"BP" (Biological process), "MF" (Molecular function) or "CC"
(Cellular Component). Cellular Component is not included with the package's
standard data and will thus yield no results.}

\item{expression_data}{expression data dataframe with at least 1 or more
rows. Some of the columns should be named certain ways and the dataframe
should contain certain information neccesary for calculation;
1st neccesary col; name; "GeneID" (EnsemblIDs should go here).
2nd neccesary col; name; "GeneType" (genetype as described in the Ensembl
database).
Next, a range of n-columns long, containing expression values, names do not
matter here. (You can use the existing dataset as an example)}

\item{ensembl_to_go_id_conversion_df}{dataframe with corresponding GOIDs and
EnsemblIDs. As well as the expression value dataframe, this dataframe should
contain certain information with specific columnnames as well;
1st neccesary col; name; "ensembl_gene_id" (Ensembl gene IDs).
2nd neccesary col; go_id; "go_id" (Gene Ontology term ID).
3rd and last neccesary col; "namespace_1003" (Toplevel GO term).
(You can use the existing dataset as an example)}

\item{start_expression_col}{should be the starting column inside
"expression_data" containing expression values. default is 4, catered to
the included dataset.}

\item{end_expression_col}{should be the last column inside "expression_data"
containing expression values. default is 22, catered to the included
dataset.}

\item{enrichment_cutoff}{cutoff number for the amount of genes to be
enriched in the enrichment analysis. (default is 250)}

\item{method}{which statistical method to use for the prediction, currently
there are 5 available; "pearson", "spearman", "kendall", "fisher", "sobolev"
and "combine".}
}
\value{
The resulting dataframe with prediction of similar GO terms.
These are ordered with respect to FDR values. The following columns will be
in the dataframe;
GOID - Gene Ontology ID,
Ontology - Ontology type (MF or BP),
FDR - False Positive Rate,
Term - description of GOID,
used_method - the used method to determine the ontology term similarity
}
\description{
Predicts annotation of un-annotated genes based on existing
Gene Ontology annotation data and correlated expression patterns.
}
\details{
This function is specifically made for predicting lncRNA annotation by
assuming "guilt by association". For instance, the expression data in this
package is actually based on mRNA expression data, but correlated with
lncRNA. This expression data is the used in combination with mRNA GO
annotation to calculate similarity scores between GO terms,
}
\examples{
# Example with default dataset, take a look at the data documentation
# to fully grasp what's going on.
GAPGOM::expression_prediction_function('ENSG00000228630', 'BP')

}
