% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_scoring.R
\name{expression_semantic_scoring}
\alias{expression_semantic_scoring}
\title{GAPGOM - expression_prediction()}
\usage{
expression_semantic_scoring(gene_id, expression_set, method = "combine")
}
\arguments{
\item{gene_id}{gene rowname to be compared to the other GO terms.}

\item{expression_set}{ExpressionSet class containing expression values and
other useful information, see GAPGOM::f5_example_data
documentation for further explanation of this type. If you want a custom
ExpressionSet you have to define one yourself.}

\item{method}{which statistical method to use for the prediction, currently
there are 5 available; "pearson", "spearman", "kendall", "fisher", "sobolev"
and "combine".}
}
\value{
The resulting dataframe with prediction of similar GO terms.
These are ordered with respect to FDR values. The following columns will be
in the dataframe;
GOID - Gene Ontology ID,
Ontology - Ontology type (MF or BP),
FDR - False Positive Rate,
Term - description of GOID,
used_method - the used method to determine the ontology term similarity
}
\description{
Predicts annotation of un-annotated genes based on existing
Gene Ontology annotation data and correlated expression patterns.
}
\details{
This function is specifically made for predicting lncRNA annotation by
assuming "guilt by association". For instance, the expression data in this
package is actually based on mRNA expression data, but correlated with
lncRNA. This expression data is the used in combination with mRNA GO
annotation to calculate similarity scores between GO terms,
}
\examples{
# Example with default dataset, take a look at the data documentation
# to fully grasp what's going on with making of the filter etc. (Biobase
# ExpressionSet)

# set an artbitrary gene you want to find similarities for. (5th row in this
# case)
gid <- "ENSG00000228630"
result <- GAPGOM::expression_semantic_scoring(gid,
                                              GAPGOM::expset)

}
