% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data_preparation.R
\name{fantom_download}
\alias{fantom_download}
\title{GAPGOM - fantom_download()}
\usage{
fantom_download(organism = "human", unpack = TRUE, noprompt = FALSE)
}
\arguments{
\item{organism}{Either "mouse" or "human". FANTOM5 only has these two as
fully annotated+tpm normalized datasets.}

\item{unpack}{Default=TRUE, if set to TRUE, file will be unpacked 
automatically.}

\item{noprompt}{Default=FALSE, user prompt, if you are 100% sure you want to
download and have enough space, set this to TRUE.}
}
\value{
The resulting filename/location of the file or NULL if cancelled.
}
\description{
Downloads and unpacks fantom5 data of either human or mouse.
}
\details{
This function downloads the whole fantom5 dataset and unpacks it. automatic
unpacking can be turned off. The file is downloaded to a special caching
directory, which will be returned on exit.
}
\examples{
fantom_file <- fantom_download(organism = "mouse", noprompt = TRUE)

}
