% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{fantom_filter_entrez}
\alias{fantom_filter_entrez}
\title{GAPGOM internal - fantom_filter_entrez()}
\usage{
fantom_filter_entrez(fan)
}
\description{
filter on entrez id with raw fantom5 matrix.
}
\section{Notes}{

This function is an internal function and should not be called by the user.
}

