% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preperation.R
\name{fantom_load_expressionset}
\alias{fantom_load_expressionset}
\title{convert into expressionset}
\usage{
fantom_load_expressionset(filepath, colselector, verbose = F)
}
\description{
convert into expressionset
}
