% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preperation.R
\name{fantom_load_raw}
\alias{fantom_load_raw}
\title{fantom5 data load.}
\usage{
fantom_load_raw(filepath, verbose = F)
}
\description{
fantom5 data load.
}
\examples{
ft5_r <- fantom_load_raw("/media/casper/USB_ccpeters/internship_thesis/data/hg19.cage_peak_phase1and2combined_counts.osc.txt2", verbose=T)
}
