% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrical_measures.R
\name{Geometrical_measures}
\alias{Geometrical_measures}
\alias{sobolev_metric}
\alias{fisher_metric}
\title{UGenePred - Geometrical Similarity Measures}
\usage{
sobolev_metric(x, y)

fisher_metric(x, y)
}
\arguments{
\item{x}{gene expression value vector (numeric) of reference/novel gene
that needs to be compared to the dataset.}

\item{y}{also a gene expression value vector, the idea here is that indexes
between x and y correspond for same or similar tissues for accurate
similarity. x and y may also be swapped.}
}
\value{
Hard to have one return section for all functions,
might want to have a manual list here.
}
\description{
These geometric functions will calculate the distance between two random
sets of variables. In our usecase this would be an expression values (FPKM).
Given two gene expression value lists, semantic similarity will be
calculated between them. This is then returned.
}
\details{
Fisher metric is based on G. Lebanon et al.'s implementation. [1]

The sobolev metrix is based on T. Villmann et al's implementation. [2]
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value
sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

\section{2. testffdklmvl}{
NA
}

\references{
[1] Villmann T: \strong{Sobolev metrixs for learning of functional
data - mathematical and theoretical aspects.} In:
\emph{Machine Learning Reports}.
Edited by Villmann T, Schleif F-m, vol. 1. Leipzig, Germany: Medical
Department, University of Leipzig; 2007: 1-13.

[2] Lebanon G: \strong{Learning riemannian metrics.} In:
\emph{Proceedings of the Nineteenth conference on Uncertainty in Artificial
Intelligence; Acapulco Mexico.} Morgan Kaufmann Publishers Inc. 2003: 362-369
}
