% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_checks.R
\name{glb_checks}
\alias{glb_checks}
\alias{.check_ifclass}
\alias{.check_gos}
\alias{.check_custom_gene}
\alias{.check_idtype}
\alias{.check_organism}
\alias{.check_ontology}
\alias{.check_method}
\title{GAPGOM internal - Global argument checks}
\usage{
.check_ifclass(obj, classname, objname, accept_null = TRUE)

.check_gos(gos)

.check_custom_gene(cus_gen)

.check_idtype(idtype, organism)

.check_organism(organism)

.check_ontology(ontology)

.check_method(method)
}
\value{
output is different on a case-to-case basis, but is a boolean 
determining if the input is correct.
}
\description{
These functions are internal functions and should not be called by the user.
}
\details{
These check make sure that user input is correct for both TopoICSim and 
lncRNApred. They mainly include very generalistic checks.
}
\section{Notes}{

Internal functions used in both algorithms.
}

\keyword{internal}
