% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_translation_df.R
\docType{data}
\name{id_translation_df}
\alias{id_translation_df}
\title{GAPGOM - id_translation_df}
\format{An translation dataframe between ensembl ids and goids.
\describe{
  \item{ORIGID}{original id as in the expset, ensembl.}
  \item{GO}{GOID term}
}}
\source{
\url{https://tare.medisin.ntnu.no/pred_lncRNA/}
}
\usage{
id_translation_df
}
\description{
An translation dataframe between ensembl ids and goids. This dataframe 
belongs with the expset data.
}
\details{
https://tare.medisin.ntnu.no/pred_lncRNA/
}
\keyword{datasets}
