% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{misc_functions}
\alias{misc_functions}
\alias{misc_pearson}
\alias{misc_sobolev}
\alias{misc_fisher}
\alias{misc_kendall}
\alias{misc_spearman}
\title{GAPGOM internal - Miscellaneous functions}
\usage{
misc_pearson(x, args)

misc_sobolev(x, args)

misc_fisher(x, args)

misc_kendall(x, args)

misc_spearman(x, args)
}
\arguments{
\item{x}{vector of expression values}

\item{args}{internal arguments, used for target expression datas}
}
\description{
These functions are misc. functions that are mostly used in applies. They
are defined here probably because they are used often.
}
\section{Notes}{

These functions are internal functions and should not be called by the user.
}

\keyword{internal}
