% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_prediction.R
\name{prediction_function}
\alias{prediction_function}
\title{UGenePred::prediction_function("ENSG00000228630", "BP", ExpressionData, EnsemblID2GOID, 4, ncol(ExpressionData))}
\usage{
prediction_function(gene_id, ontology, expression_data,
  ensembl_to_go_id_conversion_df, start_expression_col, end_expression_col,
  enrichment_cutoff = 250, method = "combine")
}
\description{
UGenePred::prediction_function("ENSG00000228630", "BP", ExpressionData, EnsemblID2GOID, 4, ncol(ExpressionData))
}
