% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing_functions.R
\name{prepare_score_df}
\alias{prepare_score_df}
\title{GAPGOM internal - prepare_score_df()}
\usage{
prepare_score_df(pc_ensmbl_id, score, gene_id)
}
\arguments{
\item{score}{score vector}

\item{gene_id}{Gene ID originally selected in the prediction function.}

\item{pc_ensmlb_id}{Ensmlb ID's for protein coding expression data.}
}
\value{
The score dataframe with ensmbl ID's
}
\description{
This function is an internal function and should not be called by the user.
}
\details{
Prepared score dataframe for the enrichment analysis.
}
\section{Notes}{

internal function used in prediction_function()
}

