% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preperation.R
\name{set_go_data}
\alias{set_go_data}
\title{GAPGOM internal - set_go_data()}
\usage{
set_go_data(organism, ontology)
}
\description{
Set GO data (this function purely makes choosing Bioconductor datasets a 
little easier)
}
\section{Notes}{

This function is an internal function and should not be called by the user.
}

