% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{topo_ic_sim}
\alias{topo_ic_sim}
\title{Calculate topo_ic_sim between two gene lists}
\usage{
topo_ic_sim(gene_list1, gene_list2, ont = "MF", organism = "human",
  drop = NULL)
}
\description{
Calculate topo_ic_sim between two gene lists
}
