% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{topo_ic_sim_g1g2}
\alias{topo_ic_sim_g1g2}
\title{GAPGOM - topo_ic_sim_g1g2()}
\usage{
topo_ic_sim_g1g2(gene1, gene2, ontology = "MF", organism = "yeast",
  progress_bar = F, garbage_collection = F, go_data = NULL,
  drop = NULL, translation_to_goids = NULL, all_go_pairs = NULL)
}
\arguments{
\item{gene1}{Gene ID of the first Gene.}

\item{gene2}{Gene ID of the second Gene.}

\item{ontology}{desired ontology to use for similarity calculations.
One of three;
"BP" (Biological process),
"MF" (Molecular function) or
"CC" (Cellular Component).}

\item{organism}{organism where to be scanned genes reside in, this option
is neccesary to select the correct GO DAG. Options are based on the org.db
bioconductor package;
http://www.bioconductor.org/packages/release/BiocViews.html#___OrgDb
Following options are available: "fly", "mouse", "rat", "yeast",
"zebrafish", "worm", "arabidopsis", "ecolik12", "bovine", "canine",
"anopheles", "ecsakai", "chicken", "chimp", "malaria", "rhesus", "pig",
"xenopus".}

\item{progress_bar}{Whether to show the progress of the calculation 
(default = FALSE)}

\item{garbage_collection}{whether to do R garbage collection. This is
useful for very large calculations/datasets, as it might decrease ram usage.
This option might however increase calculation time.}

\item{go_data}{correct GO data from specefic species.}

\item{drop}{vector of GOID you want to exclude from the analysis.}

\item{all_go_pairs}{dataframe of GO Term pairs with a column
representing similarity between the two. it is recommended to always keep
this on default unless you know what you are doing.}
}
\value{
List containing the following;
$GeneSim;
similarity between genes taken from the mean of all term 
similarities.
$GO1;
vector of GO's from first gene
$GO2;
vector of GO's from second gene
$AllGoPairs;
All possible GO combinations with their semantic distances (matrix)
}
\description{
Algorithm to calculate similarity between GO terms of two genes.
}
\details{
This function is made for calculating topological similarity of two Genes
given their GO terms in the GO DAG structure. The topological similarity is
based on edge weights and information content (IC). [1]
}
\examples{
GAPGOM::topo_ic_sim_g1g2("218", "501", ont="MF", organism="human", drop=NULL)

}
\references{
[1] Ehsani R, Drablos F: \strong{TopoICSim: a new semantic
similarity measure based on gene ontology.} \emph{BMC Bioinformatics} 2016,
\strong{17}(1):296)
}
