% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{topo_ic_sim_term}
\alias{topo_ic_sim_term}
\title{GAPGOM - topo_ic_sim_term()}
\usage{
topo_ic_sim_term(organism, ontology, go1, go2, go_data = NULL)
}
\arguments{
\item{organism}{where to be scanned genes reside in, this option
is neccesary to select the correct GO DAG. Options are based on the org.db
bioconductor package;
http://www.bioconductor.org/packages/release/BiocViews.html#___OrgDb
Following options are available: "fly", "mouse", "rat", "yeast",
"zebrafish", "worm", "arabidopsis", "ecolik12", "bovine", "canine",
"anopheles", "ecsakai", "chicken", "chimp", "malaria", "rhesus", "pig",
"xenopus".}

\item{ontology}{desired ontology to use for similarity calculations.
One of three;
"BP" (Biological process),
"MF" (Molecular function) or
"CC" (Cellular Component).}

\item{go1}{GO term of first term.}

\item{go2}{GO term of second term.}

\item{go_data}{prepared go_data, from the set_go_data function. It is
practically the same as in GOSemSim, but with a slightly nicer interface.}
}
\value{
TopoICSim score between the two terms.
}
\description{
Algorithm to calculate similarity between two GO terms.
}
\details{
This function is made for calculating topological similarity of two GO
terms in the GO DAG structure. The topological similarity is based on
edge weights and information content (IC). [1]
}
\examples{
result <- topo_ic_sim_term("human", "MF", "GO:0018478", "GO:0047105")

}
\references{
[1] Ehsani R, Drablos F: \strong{TopoICSim: a new semantic
similarity measure based on gene ontology.} \emph{BMC Bioinformatics} 2016,
\strong{17}(1):296)
}
