% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{topo_ic_sim_ti_tj}
\alias{topo_ic_sim_ti_tj}
\title{topo_ic_sim algorithm to calculate similarity between two GO terms.}
\usage{
topo_ic_sim_ti_tj(go_id1, go_id2, ont, organism, weighted_dag,
  go_annotation, root, IC, all_info_go_pairs = data.frame(GO_1 = 0, GO_2
  = 0, S_1 = 0))
}
\description{
topo_ic_sim algorithm to calculate similarity between two GO terms.
}
