% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_prediction.R
\name{expression_prediction_function}
\alias{expression_prediction_function}
\title{GAPGOM - expression_prediction_function()}
\usage{
expression_prediction_function(gene_id, expression_set, id_select_vector,
  organism, ontology, enrichment_cutoff = 250, method = "combine",
  significance = 0.05, filter_pvals = FALSE, ncluster = 1)
}
\arguments{
\item{gene_id}{gene rowname to be compared to the other GO terms.}

\item{expression_set}{ExpressionSet class containing expression values and
other useful information, see GAPGOM::f5_example_data
documentation for further explanation of this type. If you want a custom
ExpressionSet you have to define one yourself.}

\item{id_select_vector}{gene rowname(s) that you want filtered out of the
dataset. For example, let's say you need to only include protein coding
genes. You then select all other genes that aren't and put the ids in this
vector.}

\item{organism}{where to be scanned genes reside in, this option
is neccesary to select the correct GO DAG. Options are based on the org.db
bioconductor package;
http://www.bioconductor.org/packages/release/BiocViews.html#___OrgDb
Following options are available: "fly", "mouse", "rat", "yeast",
"zebrafish", "worm", "arabidopsis", "ecolik12", "bovine", "canine",
"anopheles", "ecsakai", "chicken", "chimp", "malaria", "rhesus", "pig",
"xenopus". Fantom5 data only has "human" and "mouse" available depending
on the dataset.}

\item{ontology}{desired ontology to use for prediction. One of three;
"BP" (Biological process), "MF" (Molecular function) or "CC"
(Cellular Component). Cellular Component is not included with the package's
standard data and will thus yield no results.}

\item{enrichment_cutoff}{cutoff number for the amount of genes to be
enriched in the enrichment analysis. (default is 250)}

\item{method}{which statistical method to use for the prediction, currently
there are 5 available; "pearson", "spearman", "kendall", "fisher", "sobolev"
and "combine".}

\item{significance}{normalized p-values (fdr) that are below this number 
will be kept. has to be a float/double between 0-1. Default is 0.05}

\item{filter_pvals}{filters pvalues that are equal to 0 (Default=FALSE).}
}
\value{
The resulting dataframe with prediction of similar GO terms.
These are ordered with respect to FDR values. The following columns will be
in the dataframe;
GOID - Gene Ontology ID,
Ontology - Ontology type (MF or BP),
FDR - False Positive Rate,
Term - description of GOID,
used_method - the used method to determine the ontology term similarity
}
\description{
Predicts annotation of un-annotated genes based on existing
Gene Ontology annotation data and correlated expression patterns.
}
\details{
This function is specifically made for predicting lncRNA annotation by
assuming "guilt by association". For instance, the expression data in this
package is actually based on mRNA expression data, but correlated with
lncRNA. This expression data is the used in combination with mRNA GO
annotation to calculate similarity scores between GO terms,
}
\examples{
# Example with default dataset, take a look at the data documentation
# to fully grasp what's going on with making of the filter etc. (Biobase 
# ExpressionSet)

# make a filter of all the ids that contain a 0 in their expression row.
sort_list <- rownames(GAPGOM::ft5_example_data@assayData$exprs[
apply(GAPGOM::ft5_example_data@assayData$exprs, 1, 
function(row) {all(row==0)}),])
# set an artbitrary gene you want to find similarities for. (5th row in this
# case)
gid <- rownames(ft5_example_data@assayData$exprs)[5]
GAPGOM::expression_prediction_function(gid, 
                                       GAPGOM::ft5_example_data, 
                                       sort_list, 
                                       "mouse", 
                                       "BP"
                                       )
}
