% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data_preparation.R
\name{fantom_download}
\alias{fantom_download}
\title{GAPGOM - fantom_download()}
\usage{
fantom_download(down_dir, organism = "human", unpack = T,
  noprompt = F)
}
\arguments{
\item{down_dir}{The folder/directory to download the fantom5 file to.}

\item{organism}{Either "mouse" or "human". FANTOM5 only has these two as
fully annotated+tpm normalized datasets.}

\item{unpack}{Default=TRUE, if set to TRUE, file will be unpacked 
automatically.}

\item{noprompt}{Default=FALSE, user prompt, if you are 100% sure you want to
download and have enough space, set this to TRUE.}
}
\value{
The resulting filename/location of the file or NULL if cancelled.
}
\description{
Downloads and unpacks fantom5 data of either human or mouse.
}
\details{
This function downloads the whole fantom5 dataset and unpacks it. automatic
unpacking can be turned off
}
\examples{
\dontrun{
fantom_file <- fantom_download("./", organism = "mouse", noprompt = T)
}
}
