% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_data_preparation.R
\name{fantom_load_raw}
\alias{fantom_load_raw}
\title{GAPGOM - fantom_load_raw()}
\usage{
fantom_load_raw(filepath, verbose = FALSE)
}
\arguments{
\item{filepath}{filename of fantom5 file.}

\item{verbose}{Switch to TRUE for extra messages. Default=FALSE}
}
\value{
The resulting datatable containing raw fantom5 data. (Most of the
time very large!)
}
\description{
Loads raw fantom5 data from file.
}
\details{
This function loads raw fantom5 data and returns the resulting data.table/
data.frame object.
}
\examples{
fantom_file <- fantom_download(organism = "mouse", noprompt = TRUE)
ft5 <- fantom_load_raw(fantom_file, verbose=TRUE)
}
