% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preperation.R
\name{fantom_load_raw}
\alias{fantom_load_raw}
\title{strsplit(ft5_ranncom$table[!(is.na(ft5_ranncom$table$`entrezgene (genes) id associated with the transcript`) | ft5_ranncom$table$`entrezgene (genes) id associated with the transcript`==""),][1,5][[1]], "\\:")[[1]][2]
fantom5 data load.}
\usage{
fantom_load_raw(filepath, verbose = F)
}
\description{
strsplit(ft5_ranncom$table[!(is.na(ft5_ranncom$table$`entrezgene (genes) id associated with the transcript`) | ft5_ranncom$table$`entrezgene (genes) id associated with the transcript`==""),][1,5][[1]], "\\:")[[1]][2]
fantom5 data load.
}
\examples{
ft5_r <- fantom_load_raw("/media/casper/USB_ccpeters/internship_thesis/data/hg19.cage_peak_phase1and2combined_counts.osc.txt2", verbose=T)
}
