% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_prediction.R
\name{prediction_function}
\alias{prediction_function}
\title{UGenePred - prediction_function()}
\usage{
prediction_function(gene_id, ontology,
  expression_data = UGenePred::expression_data,
  ensembl_to_go_id_conversion_df = UGenePred::ensembl_id_to_go_id,
  start_expression_col = 4, end_expression_col = 22,
  enrichment_cutoff = 250, method = "combine")
}
\arguments{
\item{gene_id}{gene to be compared to the other GO terms.}

\item{ontology}{desired ontology to use for prediction}

\item{expression_data}{lncrna2function expression data matrix}

\item{ensembl_to_go_id_conversion_df}{matrix with corresponding GOIDs and EnsemblIDs.}

\item{start_expression_col}{should be the starting column containing expression values.}

\item{end_expression_col}{should be the last column containing expression values.}

\item{enrichment_cutoff}{default is 250}

\item{method}{default is 'combine'}
}
\value{
The resulting matrix with prediction of similar GO terms.
}
\description{
Predicts functions of un-annotated genes based on Gene Ontology and correlated expression patterns.
}
\examples{
expression_data <- read.table('/media/casper/USB_ccpeters/internship_thesis/papers/pred_lncRNA/lncRNA2function_data.txt', sep='\\t', head=TRUE)
ensembl_id_to_goid <- read.table('/media/casper/USB_ccpeters/internship_thesis/papers/pred_lncRNA/EG2GO.txt', sep='\\t',head=TRUE)
UGenePred::prediction_function('ENSG00000228630', 'BP', expression_data, ensembl_id_to_goid, 4, ncol(ExpressionData))

}
