% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{set_go_data}
\alias{set_go_data}
\title{GAPGOM - set_go_data()}
\usage{
set_go_data(organism, ontology, computeIC = TRUE, keytype = "ENTREZID")
}
\arguments{
\item{organism}{where to be scanned genes reside in, this option
is neccesary to select the correct GO DAG. Options are based on the org.db
bioconductor package;
http://www.bioconductor.org/packages/release/BiocViews.html#___OrgDb
Following options are available: "fly", "mouse", "rat", "yeast",
"zebrafish", "worm", "arabidopsis", "ecolik12", "bovine", "canine",
"anopheles", "ecsakai", "chicken", "chimp", "malaria", "rhesus", "pig",
"xenopus". Fantom5 data only has "human" and "mouse" available depending
on the dataset.}

\item{ontology}{desired ontology to use for prediction. One of three;
"BP" (Biological process), "MF" (Molecular function) or "CC"
(Cellular Component). Cellular Component is not included with the package's
standard data and will thus yield no results.}

\item{computeIC}{whether to compute Information Content.}

\item{keytype}{keytype used in querying of godata}
}
\value{
return godata as from GoSemSim
}
\description{
Sets GO data like GOSemSim (this function purely makes choosing datasets a 
little easier and prints available keytypes if specified incorrectly.)
}
\section{Notes}{

Internal function used in multiple functions of topoICSim.
}

