% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{topo_ic_sim_g1g2}
\alias{topo_ic_sim_g1g2}
\title{GAPGOM - topo_ic_sim_g1g2()}
\usage{
topo_ic_sim_g1g2(gene1, gene2, go_data, ontology = "MF",
  organism = "yeast", drop = NULL,
  all_info_go_pairs = data.frame(GO_1 = 0, GO_2 = 0, S_1 = 0))
}
\arguments{
\item{gene1}{Gene ID of the first Gene.}

\item{gene2}{Gene ID of the second Gene.}

\item{go_data}{correct GO data from specefic species.}

\item{ontology}{desired ontology to use for similarity calculations.
One of three;
"BP" (Biological process),
"MF" (Molecular function) or
"CC" (Cellular Component).}

\item{organism}{organism where to be scanned genes reside in, this option
is neccesary to select the correct GO DAG. Options are based on the org.db
bioconductor package;
http://www.bioconductor.org/packages/release/BiocViews.html#___OrgDb
Following options are available: "fly", "mouse", "rat", "yeast",
"zebrafish", "worm", "arabidopsis", "ecolik12", "bovine", "canine",
"anopheles", "ecsakai", "chicken", "chimp", "malaria", "rhesus", "pig",
"xenopus".}

\item{drop}{vector of GOID you want to exclude from the analysis.}

\item{all_info_go_pairs}{dataframe of GO Term pairs with a column
representing similarity between the two. it is recommended to always keep
this on default unless you know what you are doing.}
}
\value{
similarity between genes taken from the mean of all term 
similarities.
}
\description{
Algorithm to calculate similarity between GO terms of two genes.
}
\details{
This function is made for calculating topological similarity of two Genes
given their GO terms in the GO DAG structure. The topological similarity is
based on edge weights and information content (IC). [1]
}
\examples{
GAPGOM::topo_ic_sim_g1g2("218", "501", ont="MF", organism="human", drop=NULL)

}
\references{
[1] Ehsani R, Drablos F: \strong{TopoICSim: a new semantic
similarity measure based on gene ontology.} \emph{BMC Bioinformatics} 2016,
\strong{17}(1):296)
}
