% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoicsim_algorithms.R
\name{topo_ic_sim_g1g2}
\alias{topo_ic_sim_g1g2}
\title{topo_ic_sim algorithm to calculate similarity between two genes or gene products.}
\usage{
topo_ic_sim_g1g2(gene1, gene2, go_data, ont = "MF", organism = "yeast",
  drop = NULL, all_info_go_pairs = data.frame(GO_1 = 0, GO_2 = 0, S_1 =
  0))
}
\description{
topo_ic_sim algorithm to calculate similarity between two genes or gene products.
}
